+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Configuration overrides for the JVM security properties, such as DNS caching ([#24]).
+
 ### Changed
 
 - Upgraded to EDC 0.1.2 ([#23]).
@@ -14,3 +18,4 @@ All notable changes to this project will be documented in this file.
 [#20]: https://github.com/stackabletech/edc-operator/pull/20
 [#21]: https://github.com/stackabletech/edc-operator/pull/21
 [#23]: https://github.com/stackabletech/edc-operator/pull/23
+[#24]: https://github.com/stackabletech/edc-operator/pull/24